# Don't expand vector if there is nothing left to write into it.