A common C++ pattern is as follows:

vector v;
v.reserve(size);
// add size elements to v in some way, without v being required to expand

In Rust, users will equivalently do the following:
let mut v = Vec::with_capacity(size);
// add size elements to v in some way, without v being required to expand

The method read_to_end violates this expectation, by increasing the amount of space the vector is taking if the buffer being currently read in is exactly the right size required.

This results in a large performance drop in these cases- I attach benchmarking code and results.  This PR fixes this, so that if read_to_end is called with exactly the correct number of bytes allocated, the vector will not expand in size.

    #![feature(test)]
    extern crate test;
    use std::io::Read;

    #[bench]
    fn read_to_end_1024(b: &mut test::Bencher) {
        let buf = [1u8; 1024];
        let mut src = &buf[..];
        let mut dst = Vec::with_capacity(1024);
        b.iter(|| {
            test::black_box(src.read_to_end(&mut dst))
        });
    }

    #[bench]
    fn read_to_end_64k(b: &mut test::Bencher) {
        let buf = [1u8; 1024 * 64];
        let mut src = &buf[..];
        let mut dst = Vec::with_capacity(64);
        b.iter(|| {
            test::black_box(src.read_to_end(&mut dst))
        });
    }

    #[bench]
    fn read_to_end_normal_test(b: &mut test::Bencher) {
        let buf = [1u8; 100000];
        let mut src = &buf[..];
        let mut dst = Vec::new();
        b.iter(|| {
            test::black_box(src.read_to_end(&mut dst))
        });
    }

    #[bench]
    fn read_to_end_small_test(b: &mut test::Bencher) {
        let buf = [1u8; 1000];
        let mut src = &buf[..];
        let mut dst = Vec::new();
        b.iter(|| {
            test::black_box(src.read_to_end(&mut dst))
        });
    }

    #[test]
    fn read_to_end_doesnt_expand_capacity_unless_necessary() {
        const BUF_SIZE: usize = 1024;
        let buf = [1u8; BUF_SIZE];
        let mut src = &buf[..];
        let mut dst = Vec::with_capacity(BUF_SIZE);
        let init_capacity = dst.capacity();
        assert_eq!(src.read_to_end(&mut dst).unwrap(), BUF_SIZE);
        assert_eq!(init_capacity, dst.capacity());
    }

    /*
    Results on my machine:

    Without Modification:

    test read_to_end_1024        ... bench:      9972 ns/iter (+/- 2666)
    test read_to_end_64k         ... bench:   2568395 ns/iter (+/- 793768)
    test read_to_end_normal_test ... bench:   2133921 ns/iter (+/- 554167)
    test read_to_end_small_test  ... bench:     10703 ns/iter (+/- 2967)


    With Modification:

    test read_to_end_1024            ... bench:       750 ns/iter (+/- 30)
    test read_to_end_64k              ... bench:     74184 ns/iter (+/- 14296)
    test read_to_end_normal_test  ... bench:   1525546 ns/iter (+/- 200031)
    test read_to_end_small_test   ... bench:      5651 ns/iter (+/- 1128)
    */