@alexcrichton 2015-04-03
In the benchmarks the same vector is being pushed onto on each iteration of the loop, and the 64k benchmark doesn't reserve 64k of space (it just reserves 64 bytes), were these both intentional? I tried tweaking both of these and got these [before and after numbers](https://gist.github.com/alexcrichton/31122565e893d6599453).

I'm somewhat wary of optimizing for microbenchmarks like this, I feel like the implementation of these common functions should be expressible in concise terms, but this `read_to_end` function continues to become more sprawling over time. This also issues 2x more syscalls when the reader is backed by something like a `File` or `TcpStream`, which may or may not be desirable. 

So I suppose I have a few questions:

* Can you measure how this compares when backed by a syscall-oriented reader?
* Would it be possible to make this more concise? The two `match` statements are very similar and are a little unsightly to have duplicated.

Thanks for continuing to press on this function!