@bcoopers 2015-04-03
>In the benchmarks the same vector is being pushed onto on each iteration of the loop, and the 64k benchmark doesn't reserve 64k of space (it just reserves 64 bytes), were these both intentional?

The first of these was not intentional, the second of these was (I wanted to do one where it continued to expand, the other where it did not. The 64 reserved at the end will continue being multiplied by 2, which means it will eventually meet the 64k requirement exactly, and thus be expanded to 128k by the original but kept at 64k by the modified version.)

> I tried tweaking both of these and got these before and after numbers.

On my computer your test gives:

Before:
test read_to_end_1024        ... bench:      7812 ns/iter (+/- 1498)
test read_to_end_64k         ... bench:   2185142 ns/iter (+/- 260208)
test read_to_end_normal_test ... bench:   2180402 ns/iter (+/- 498111)
test read_to_end_small_test  ... bench:      7876 ns/iter (+/- 1808)

After:
test read_to_end_1024        ... bench:      7936 ns/iter (+/- 3158)
test read_to_end_64k         ... bench:    147839 ns/iter (+/- 95546)
test read_to_end_normal_test ... bench:   1981639 ns/iter (+/- 343215)
test read_to_end_small_test  ... bench:       802 ns/iter (+/- 342)

I ran these a few times, they seemed pretty consistent.  You can see it's considerably faster for the largest case where the old would need to reallocate, and the new would not.  It's obvious your computer is much beefier than mine- maybe running with a larger buffer on your machine would bring out the performance difference?

> I'm somewhat wary of optimizing for microbenchmarks like this, I feel like the implementation of these common functions should be expressible in concise terms, but this read_to_end function continues to become more sprawling over time.

I would generally agree, but in this case I think it's an important benchmark, since it is a fairly common use case to treat a vector as a dynamic array rather than a variable length array, assigning it a capacity exactly what you know it will reach and then never growing it past that capacity (eg, I know the file is exactly this long, let me reserve that much up front so it never has to grow).  In this function, using a vector in this way will result in using up twice as much memory as necessary, as well as reallocating the entire vector for no reason.

Right now the only way to get optimal performance in this case is to reserve exactly one more byte than you need.  It seems better to make this function more complex rather than requiring every Rust programmer to remember to reserve the extra byte when using this function.

Given that though...

> Can you measure how this compares when backed by a syscall-oriented reader?

This is a good point.  I'll look into this, and try testing on a file.  It could be that the syscall will wipe out any performance increases.

> Would it be possible to make this more concise? The two match statements are very similar and are a little unsightly to have duplicated.

Yeah this would be nice.  I'll look into this as well.