@nikomatsakis 2015-04-03
r? @alexcrichton (changing review since Alex basically owns this code)