@bcoopers 2015-04-03
I tested it on files of different sizes, and in every case got either a large performance increase (sometimes twice as fast) or, if it was slower, was on the order of 1% or so (within the errors of the bench).  https://gist.github.com/bcoopers/c5b363a4effc4f78509b