@bcoopers 2015-04-03
Looking at it, I don't see a way to reduce the complexity of the code.  Though the results of the match are similar, what they are actually doing is entirely different.  You could combine them by having one match statement, with an if statement saying whether to write it to a temporary buffer or the end of the vector, and another one inside the results saying whether to copy the byte from the buffer over or not- but I think this would make the function look more cluttered, not less.

I'll keep thinking about it, but the solution isn't apparent to me right now.