@alexcrichton 2015-04-03
> On my computer your test gives:

Whoa, I have a feeling my box isn't 150x faster than yours! How are you compiling/running the benchmarks? I was compiling as `rustc foo.rs --test -Clto -Copt-level=3` and then running via `./foo --bench`. Did you perhaps forget to optimize?

> I would generally agree, but in this case I think it's an important benchmark ...  In this function, using a vector in this way will result in using up twice as much memory as necessary, as well as reallocating the entire vector for no reason.

Yeah I do agree with this, it does seem bad! I just want to get a handle on what sort of overhead we're talking about :)

> in every case got either a large performance increase (sometimes twice as fast)

Hm, I'm getting [quite different numbers](https://gist.github.com/7c660b0bee6dd7c82de0) myself. For me it looks like this PR is a little faster, but only slightly so. This may be related to the answer to my first question though!