@bcoopers 2015-04-03
You're absolutely right, I did not turn the optimizer on for my tests.  Sorry about that.

With the optimizer on, I get pretty different results from yours still... https://gist.github.com/bcoopers/cd2465904a3c7760c577

On my machine the change seems to hurt more at smaller read sizes (which makes sense, since the syscall stays the same price while the copy/reallocate gets more expensive as the vector gets larger) and results in maybe a 15% improvement for a large (10 megabyte) file.

I'll leave it up to you at this point, whether to merge or close it. It's not particularly clear whether this change is beneficial or not, as my results suggest it strongly depends on how much data is being read in.  It certainly wasn't as beneficial as I had hoped, or as my initial (mistaken) benchmark suggested.