@bcoopers 2015-04-03
Hmm it says I closed it, which I am pretty sure I did not... maybe I accidentally clicked something... in any case I suspect you were going to close it anyway.  Let me know if you decide differently and I can recreate the